+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## 0.0.4
+- Setup uses `pyproject.toml` with `hatchling` backend
+- Added tests and linting with `noxfile.py`
 - Created `samples` sub-package with sample price data
 - Removed data files from tests folder
 - Removed relative links from readme (except showcase.svg)