+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## In progress
+
+
 ## 0.0.4
 - Added tasks.py
 - Added github workflow